requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
dj-database-url==2.3.0
openpyxl==3.1.2
pandas==2.1.3
numpy==1.26.4
matplotlib==3.10.1
Pillow==10.0.1
reportlab==4.0.4
weasyprint==60.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
django-auditlog==2.3.0
django-bootstrap4==23.2
django-bootstrap5==23.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==4.1
django-filter==23.3
django-guardian==2.4.0
django-import-export==3.3.1
django-money==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.32.3
beautifulsoup4==4.13.3
polib==1.2.0
yfinance==0.2.55
ta==0.11.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
et_xmlfile==2.0.0
odfpy==1.4.1
tablib==3.5.0
defusedxml==0.7.1
diff-match-patch==20241021
MarkupPy==1.18
PyYAML==6.0.2